feat: Update OpStack's digestFinalizedBlock method to not store files
This commit updates the `digestFinalizedBlock` method in the `OpStack` class to not store the block content as files. Previously, the method had an option to store the content as files, but it was set to `true` by default. This change sets the option to `false`, ensuring that the content is not stored as files.