Refactor TimeSynchronizer class to add autoStart option
This commit refactors the TimeSynchronizer class to add an autoStart option. The autoStart option determines whether the sync loop should start automatically when the TimeSynchronizer is instantiated. If autoStart is true, the sync loop will start; otherwise, it will not. This change improves the flexibility and control over when the sync loop should begin.